moving writeStandaloneValidate to a different file.
